SinglePlaneDebug: output non-garbage values
The local array declared inside debug...() ends its lifetime upon return,
so by the time set_frame_props reads the memory, it can contain anything
(or more accurately, the pointer dereference is invalid and the behavior
is undefined). For example, in my tests, I get all-zero BoreAdjustment.

To ensure the correct values are actually output,
put them in a memory location that survives the function return.
It seems most natural to have the caller provide this location.